This is my version of Part One
I’m sorry; I didn’t see that you had already started Part 1. I had just bookmarked the Project list and saw that it was not in “In Progress” yet so I did it. I uploaded just because I figured I might as well. It seems to meet the criteria, but we can use whichever.