The last character doesn't seem to be coming through
So I was doing my last little checks to make sure we were following all the rules and I noticed that 4B is not printing out the last character. I am working on it, but please test if you can to see if it is happening for you with these two files.